fixed some msvc warnings

+++ src/client/CClient_Game.cpp
@@ -116,7 +116,7 @@ void CClient::Simulation()
 						(unsigned int)CLAMP(w->getPos().x, 0.0f, (float)game.gameMap()->GetWidth() - 1),
 						(unsigned int)CLAMP(w->getPos().y, 0.0f, (float)game.gameMap()->GetHeight() - 1));
 			if(mat.damage)
-				InjureWorm(w, mat.damage, -1);
+				InjureWorm(w, (float)mat.damage, -1);
 
 			// Check if this worm picked up a bonus
